Google_ADK/7-Multi-Agent_2/Travel_Planner_Multi_Agents/pyproject.toml
[project]
name = "travel-planner-multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.22.1",
    "litellm>=1.80.16",
]
